+++ src/structs/objects.h
@@ -14,13 +14,13 @@ struct Sphere {
     Vector3 position;
     float radius;
     // 16 bytes
-    int materialIndex;
+    float materialIndex;
     float _padding_1[3];
 
     Sphere(Vector3 _position, float _radius, int _materialIndex) {
         position = _position;
         radius = _radius;
-        materialIndex = _materialIndex;
+        materialIndex = (float)_materialIndex;
     }
 };
 
@@ -41,7 +41,7 @@ struct Triangle {
     Vector2 uv1;
     // 16 bytes
     Vector2 uv2;
-    int materialIndex;
+    float materialIndex;
     float _padding_4;
 
     Triangle(Vector3 _v0, Vector3 _v1, Vector3 _v2, int _materialIndex, Vector2 _uv0 = {0, 0},
@@ -49,8 +49,7 @@ struct Triangle {
         v0 = _v0;
         v1 = _v1;
         v2 = _v2;
-        materialIndex = _materialIndex;
-
+        materialIndex = (float)_materialIndex;
         uv0 = _uv0;
         uv1 = _uv1;
         uv2 = _uv2;
